Revert "Merge pull request #30 from solidusio-contrib/mm/post-process-import"
The post-processors code also introduced change in the context that
made it permanent across rows. This led to bugs due leaked information
between rows (see issue #35 for details about the errors).

The gist of this revert is the removal of call to #dup on the context.

This reverts commit 1c9685b19bc221854ef32cd6591a1fa74339d221, reversing
changes made to 644ba814521273ffd2100e333a05550b0a5acff6.