Add spec for option types in DB, not on product
This covers the case where the option types already exist inside of the
DB, but do not exist on the product that is being imported.

In this scenario, we should not be saving new option types, but with
the current code it does. I'm marking the spec as pending because it
will fail, but will be passing in the next commit.