Look for OptionType from database
In the case where the option types already exist in the DB, but not
in the product, we should look for them there first, so that we can
use the ones we've already got instead of making more.

This also adds some code so we're not overriding any information already
present on the option type, specifically the presentation field, as we
cannot really assume that the presentation should be equal to the name.