Process taxons if "Tags" or "Type" are empty
Before this commit, an empty "Tags" or "Type" cell in a CSV would cause
this processor to fail.

I believe it's reasonable to assume that some products may not include
tags on import. And, while I'd prefer all products be imported with a
"Type" personally, I don't this this is should be a hard requirement
that makes rows fail in a product import.