Check for duplicated option values and avoid adding them
When the import file has more lines for each variant in order to
specify more images per variant the mapping to the variant is made
through options instead of variant sku (which is optional). As a
result of repeating the options on all lines they are also added
by the importer multiple times ending. This change avoids adding
them if already present.