Refactor import spec
A rubocop version bump is causing the previous implementation of this
spec to fail due to "stubbing an object under test" on line 68.

I took this opportunity to refactor this spec using the following
rules:
 - Don't mock objects under test.
 - If this object is coupled to a collaborator, the SolidusImporter::Row
   class, we should not mock the associated objects, since the
   dependency needs to be tested.
 - Don't test ActiveRecord functionality.

I tried to avoid creating/saving as much as possible, to maintain speed, but
since the import model uses persistence layer functions on the
associated Row objects, creating the rows is unavoidable without
introducing a circular dependency.