Process Rows Async (in Groups)
This refactor groups rows that have related data (mostly Orders) so that they can be processed together. A huge problem in the original extension is that processing related rows for Orders, with different Line Items, required the whole context to be saved and ready post-import, which lead to huge times to process the whole import and also problems with visualizing errors.

Overall, here was the approach taken:
- each `row` now includes an `entity_id`, which determines the grouping. This is, for example, an Order number
- when processing the import, the `ProcessRow` is now wrapped by a new `ProcessRowGroupJob`
- the `after_context` method was renamed to `after_group_import`. It wasn't used for anything besides Orders anyway, and it was just a hack for importing Orders post-row processing
- a few tweaks were done to wrapping the Import - renaming methods, logic moved around - to accomodate for the fact it now finishes in a job

With this, the Importer `options` (ability to specify processor callables / import class) are lost because they cannot be passed to this job. This does impact tests but it shouldn't impact operation much. To preserve this customization, I would move the code on `ProcessRowGroupJob` to a service and pass the options there.