Fix Several spec errors
To make sure that the introduced refactor doesn't break the behavior, I had to introduce a way to handle rows in-line when testing. It'd be better to inject the callable to `ProcessImport` but that would severely increase the difficulty of refactoring the test suite.

Since several tests were breaking, I used as a bench-mark the master branch and fixed mostly what was affected by the refactor, plus a few easy fixes and rubocop warnings.